+++ tests/test_printer.py
@@ -952,43 +952,3 @@ def test_get_printed_name():
     printed = StringIO()
     picked_name = Printer(printed).print_ssa_value(val)
     assert f"%{picked_name}" == printed.getvalue()
-
-
-def test_indented():
-    output = StringIO()
-    printer = Printer(stream=output)
-    printer.print("\n{")
-    with printer.indented():
-        printer.print("\nhello\nhow are you?")
-        printer.print("\n(")
-        with printer.indented():
-            printer.print("\nfoo,")
-            printer.print("\nbar,")
-            printer.print("\n")
-            printer.print_string("test\nraw print!", indent=0)
-            printer.print_string("\ndifferent indent level", indent=4)
-        printer.print("\n)")
-    printer.print("\n}")
-    printer.print("\n[")
-    with printer.indented(amount=3):
-        printer.print("\nbaz")
-    printer.print("\n]\n")
-
-    EXPECTED = """
-{
-  hello
-  how are you?
-  (
-    foo,
-    bar,
-    test
-raw print!
-        different indent level
-  )
-}
-[
-      baz
-]
-"""
-
-    assert output.getvalue() == EXPECTED
